fix: remove file from s3 just if exists

+++ src/main/java/com/karoldm/bookstore/services/BookService.java
@@ -68,7 +68,9 @@ public class BookService {
 
         Book book = optionalBook.get();
 
-        fileStorageService.removeFileByUrl(book.getCover());
+        if(book.getCover() != null){
+            fileStorageService.removeFileByUrl(book.getCover());
+        }
 
         bookRepository.delete(book);
     }
